+++ vscode/webviews/Components/EnhancedContextSettings.tsx
@@ -15,7 +15,10 @@ import { PopupFrame } from '../Popups/Popup'
 import popupStyles from '../Popups/Popup.module.css'
 import styles from './EnhancedContextSettings.module.css'
 
-interface EnhancedContextSettingsProps {}
+interface EnhancedContextSettingsProps {
+    isOpen: boolean
+    setOpen: (open: boolean) => void
+}
 
 export function defaultEnhancedContextContext(): EnhancedContextContextT {
     return {
@@ -66,7 +69,7 @@ const ContextGroupComponent: React.FunctionComponent<{ group: ContextGroup; allG
 
     return (
         <>
-            <dt title={group.name}>
+            <dt title={group.name} className={styles.lineBreakAll}>
                 <i className="codicon codicon-folder" /> {groupName}
             </dt>
             <dd>
@@ -166,11 +169,13 @@ const ContextProviderComponent: React.FunctionComponent<{ provider: ContextProvi
     )
 }
 
-export const EnhancedContextSettings: React.FunctionComponent<EnhancedContextSettingsProps> = (): React.ReactNode => {
+export const EnhancedContextSettings: React.FunctionComponent<EnhancedContextSettingsProps> = ({
+    isOpen,
+    setOpen,
+}): React.ReactNode => {
     const events = useEnhancedContextEventHandlers()
     const context = useEnhancedContextContext()
     const [enabled, setEnabled] = React.useState<boolean>(useEnhancedContextEnabled())
-    const [isOpen, setOpen] = React.useState(false)
     const enabledChanged = React.useCallback(
         (event: any): void => {
             const shouldEnable = !!event.target?.checked
@@ -181,11 +186,18 @@ export const EnhancedContextSettings: React.FunctionComponent<EnhancedContextSet
         },
         [events, enabled]
     )
+
+    const hasOpenedBeforeKey = 'enhanced-context-settings.has-opened-before'
+    const hasOpenedBefore = localStorage.getItem(hasOpenedBeforeKey) === 'true'
+    if (isOpen && !hasOpenedBefore) {
+        localStorage.setItem(hasOpenedBeforeKey, 'true')
+    }
+
     return (
         <div className={classNames(popupStyles.popupHost)}>
             <PopupFrame
                 isOpen={isOpen}
-                onDismiss={() => setOpen(!isOpen)}
+                onDismiss={() => setOpen(false)}
                 classNames={[popupStyles.popupTrail, styles.enhancedContextSettingsPopup]}
             >
                 <div className={styles.enhancedContextInnerContainer}>
@@ -216,7 +228,7 @@ export const EnhancedContextSettings: React.FunctionComponent<EnhancedContextSet
                 title="Configure Enhanced Context"
             >
                 <i className="codicon codicon-sparkle" />
-                {/* Show this dot if the popover has never been opened: <div className={styles.glowyDot}/> */}
+                {isOpen || hasOpenedBefore ? null : <div className={styles.glowyDot} />}
             </VSCodeButton>
         </div>
     )